Can now calculate longitude when clicking on a tile. This was tested on the zoom=0 and 1 (tile 0,0) levels

+++ NetMapper/PointTypes/PointLatLon.cs
@@ -8,10 +8,10 @@ namespace NetMapper.PointTypes
 {
     internal class PointLatLon
     {
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
 
-        public PointLatLon(double latitude, double longitude)
+        public PointLatLon(float latitude, float longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
